Merge master into staging (#1724)
* Switch rabbitmq-setup to connect to boulder-rabbitmq

This followup for #1639 replaces localhost with boulder-rabbitmq when tests run rabbitmq-setup.
Also fixed log message to point to the server name, not 0.0.0.0, when notifying about trying to connect.

* Revert "Cat mariadb logs after failure"

This reverts commit b5cc545b48dd2ad062954e0d1fd9a225ecd54f13.

* Suppress noise on running unittests.

Remove -v for go install.

Don't print success after every successful command.

Remove `run` in front of `go test` to avoid printing each command.

This reduces the number of log lines taken by successful unittests in Travis by
66%, and makes failures much more visible and easy to read.

* Accept = in orphan-finder.

Also, when a certificate already exists, treat that as info, not error.

Update mock logger to allow matching by log level, and fix WFE and VA tests
correspondingly.

* Use a mock for OCSP responder test.

This reduces the use of a "real" SA in our tests, which will eventually allow us
to run them in parallel.

* expect client to use http://boulder/ for testing

https://github.com/letsencrypt/boulder/pull/1698

This followup for #1639 to use the boulder name, not 127.0.0.1, in subscriberAgreementURL in the test boulder instance.

* always use boulder-mysql, not localhost, for mysql

This followup for #1639 to always use the boulder-mysql, not localhost, when connecting to mysql database when testing.

* Fix test/setup.sh for removed golang packages (#1712)

Per https://groups.google.com/forum/#!topic/golang-announce/JMs4_CGbXWk, golang.org/x/tools/cmd/vet no longer exists. We don't need it anymore since we upgraded to Go 1.5, so remove it.

* Change panics in audit logger to log.Fatalf. (#1695)

* Change panics in audit logger to log.Fatalf.

These errors can happen if the syslog apparatus is unavailable, so are not that
out of the ordinary. log.Fatalf produces less noise and makes it easier to see
what's wrong.

* Fix imports.

* reloader: Split data/error callbacks (#1704)

The reloader package currently discriminates between file reading errors (in the reloader package) and data parsing errors (in the client code). Because of this, errors in a hot-replaced policy file may go unnoticed.

The former design, a (data, error) callback is replaced with a data callback and an error callback. This allows for uniform reporting of errors sourced from the library and errors sourced from the user-provided parser.

https://github.com/letsencrypt/boulder/pull/1704

* Fix a panic in cert-checker (#1714)

* Fix cert-checker panic

Fixes a silly panic in the cert-checker that would've caused it to fail outside of tests, also fixed the test to catch that silliness.

* Allow Must Staple in ECDSA and RSA profiles. (#1713)

* Allow Must Staple in ECDSA and RSA profiles.
* Port recent changes into config-next.

* Speed up docker build (#1716)

Make COPY and compilation the last commands in the Dockerfile so in the common case Docker will cache results of EXPOSE, WORKDIR and ENV commands. The CMD is eliminated as entrypoint.sh now defaults to start.py if no arguments are given. The patch eliminates setting MYSQL_CONTAINER in run-docker.sh and docker-compose.yaml as entrypoint.sh sets the variable on its own when calling create_db.sh.

In addition the patch passes arguments passed to run-docker.sh as arguments to the entryscript.sh in the container. This way running `./run-docker.sh ./test.sh ...` allows to execute tests locally.

* Reduce log levels in OCSP responder. (#1702)

* Reduce log levels in OCSP responder.
* Use mock log in test.
* Update upstream cfssl.

* Fix the ORDER BY/OFFSET from the nameset-backfill query (#1711)

* Refactor log package (#1717)

- Remove error signatures from log methods. This means fewer places where errcheck will show ignored errors.
- Pull in latest cfssl to be compatible with errorless log messages.
- Reduce the number of message priorities we support to just those we actually use.
- AuditNotice -> AuditInfo
- Remove InfoObject (only one use, switched to Info)
- Remove EmergencyExit and related functions in favor of panic
- Remove SyslogWriter / AuditLogger separate types in favor of a single interface, Logger, that has all the logging methods on it.
- Merge mock log into logger. This allows us to unexport the internals but still override them in the mock.
- Shorten names to be compatible with Go style: New, Set, Get, Logger, NewMock, etc.
- Use a shorter log format for stdout logs.
- Remove "... Starting" log messages. We have better information in the "Versions" message logged at startup.

Motivation: The AuditLogger / SyslogWriter distinction was confusing and exposed internals only necessary for tests. Some components accepted one type and some accepted the other. This made it hard to consistently use mock loggers in tests. Also, the unnecessarily fat interface for AuditLogger made it hard to meaningfully mock out.